+++ inlang/source-code/cross-sell/cross-sell-sherlock/src/index.ts
@@ -1,0 +1,16 @@
+import type { NodeishFilesystem } from "@lix-js/fs"
+import {
+	addRecommendationToWorkspace,
+	isInWorkspaceRecommendation,
+} from "./recommendation/recommendation.js"
+
+export async function isAdopted(
+	fs: NodeishFilesystem,
+	workingDirectory?: string
+): Promise<boolean> {
+	return isInWorkspaceRecommendation(fs, workingDirectory)
+}
+
+export async function add(fs: NodeishFilesystem, workingDirectory?: string): Promise<void> {
+	await addRecommendationToWorkspace(fs, workingDirectory)
+}