+++ inlang/source-code/editor/src/pages/@host/@owner/@repository/State.tsx
@@ -83,14 +83,20 @@ type EditorStateSchema = {
 	/**
 	 * Id to filter messages
 	 */
-	filteredId: () => string
-	setFilteredId: Setter<string>
+	filteredIds: () => string[]
+	setFilteredIds: Setter<string[]>
 
 	/**
 	 * TextSearch to filter messages
 	 */
 	textSearch: () => string
 	setTextSearch: Setter<string>
+
+	/**
+	 * Reference link to from where the user came from.
+	 */
+	refLink: () => string
+	setRefLink: Setter<string>
 
 	/**
 	 * The filesystem is not reactive, hence setFsChange to manually
@@ -206,9 +212,9 @@ export function EditorStateProvider(props: { children: JSXElement }) {
 	//set filter with search params
 	const params = new URL(document.URL).searchParams
 
-	const [filteredId, setFilteredId] = createSignal<string>((params.get("id") || "") as string)
+	const [filteredIds, setFilteredIds] = createSignal<string[]>(params.getAll("id") as string[])
 	createEffect(() => {
-		setSearchParams({ key: "id", value: filteredId() })
+		setSearchParams({ key: "id", value: filteredIds() })
 	})
 
 	const [textSearch, setTextSearch] = createSignal<string>((params.get("search") || "") as string)
@@ -228,6 +234,11 @@ export function EditorStateProvider(props: { children: JSXElement }) {
 	>(params.getAll("lint") as MessageLintRule["id"][])
 	createEffect(() => {
 		setSearchParams({ key: "lint", value: filteredMessageLintRules() })
+	})
+
+	const [refLink, setRefLink] = createSignal<string>((params.get("ref") || "") as string)
+	createEffect(() => {
+		setSearchParams({ key: "ref", value: refLink() })
 	})
 
 	const [localStorage] = useLocalStorage() ?? []
@@ -319,7 +330,10 @@ export function EditorStateProvider(props: { children: JSXElement }) {
 			if (args.repo?.nodeishFs === undefined) return []
 			const projects = await listProjects(args.repo?.nodeishFs, "/")
 
-			if (searchParams().project && projects.some((project) => project.projectPath === searchParams().project)) {
+			if (
+				searchParams().project &&
+				projects.some((project) => project.projectPath === searchParams().project)
+			) {
 				setActiveProject(searchParams().project)
 			} else if (projects.length === 1) {
 				setActiveProject(projects[0]?.projectPath)
@@ -529,10 +543,12 @@ export function EditorStateProvider(props: { children: JSXElement }) {
 					githubRepositoryInformation,
 					routeParams,
 					searchParams,
-					filteredId,
-					setFilteredId,
+					filteredIds,
+					setFilteredIds,
 					textSearch,
 					setTextSearch,
+					refLink,
+					setRefLink,
 					fsChange,
 					setFsChange,
 					project,
@@ -629,7 +645,7 @@ export async function pushChanges(args: {
 				name: args.user.username,
 				email: args.user.email,
 			},
-			message: "inlang: update translations",
+			message: "Fink 🐦: update translations",
 		})
 	}