+++ inlang/source-code/paraglide/paraglide-js-adapter-next/src/app/index.server.tsx
@@ -1,0 +1,2 @@
+export { default as LanguageProvider } from "./providers/LanguageProvider.js"
+export { createI18n } from "./createI18n.server.js"