+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/example/src/routes/some-subpage/+server.js
@@ -1,9 +1,0 @@
-import { base } from "$app/paths"
-import { i18n } from "$lib/i18n"
-import { redirect } from "@sveltejs/kit"
-
-export const prerender = true
-
-export function GET() {
-	redirect(303, i18n.resolveRoute(base + "/about", "fr"))
-}