+++ inlang/source-code/rpc/src/functions/index.ts
@@ -1,7 +1,6 @@
 import { subscribeNewsletter } from "./subscribeNewsletter.js"
 import { subscribeCategory } from "./subscribeCategory.js"
 import { machineTranslateMessage } from "./machineTranslateMessage.js"
-import { getNumberOfProjects } from "./getNumberOfProjects.js"
 
 export const allRpcs = {
 	machineTranslateMessage,
@@ -11,7 +10,6 @@ export const allRpcs = {
 	machineTranslate: () => undefined,
 	subscribeNewsletter,
 	subscribeCategory,
-	getNumberOfProjects,
 }
 
 export type AllRpcs = typeof allRpcs