+++ inlang/source-code/github-lint-action/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @inlang/github-lint-action
 
+## 0.2.0
+
+### Minor Changes
+
+- 8ba15bfa0: First version of Ninja-i18n
+
+## 0.1.8
+
+### Patch Changes
+
+- Retrigger release
+
+## 0.1.7
+
+### Patch Changes
+
+- Retrigger release
+
+## 0.1.6
+
+### Patch Changes
+
+- @inlang/sdk@0.26.5
+
+## 0.1.5
+
+### Patch Changes
+
+- @inlang/sdk@0.26.4
+
 ## 0.1.4
 
 ### Patch Changes