+++ inlang/source-code/paraglide/paraglide-js-adapter-vite/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @inlang/paraglide-js-adapter-vite
 
+## 1.2.18
+
+### Patch Changes
+
+- @inlang/paraglide-js-adapter-unplugin@1.4.7
+
+## 1.2.17
+
+### Patch Changes
+
+- @inlang/paraglide-js-adapter-unplugin@1.4.6
+
+## 1.2.16
+
+### Patch Changes
+
+- @inlang/paraglide-js-adapter-unplugin@1.4.5
+
 ## 1.2.15
 
 ### Patch Changes